+++ gd-core/src/tests/components/d3/TestStackedBarChart.jsx
@@ -45,7 +45,8 @@ const useStyle = makeStyles({
 const TestStackedBarChart = () => {
     const classes = useStyle();
 
-    const [container, containerRect] = useElementRect();
+    const container = React.useRef();
+    const containerRect = useElementRect(container);
 
     const tooltipContainerRef = React.useRef(null);
 
@@ -62,13 +63,13 @@ const TestStackedBarChart = () => {
                     scale: scaleBand().domain(FIXTURE.data.map(({ year }) => year)),
                     key: 'year',
                     title: 'Year',
-                    titlePadding: 50
+                    titleYPadding: 50
                 }}
                 yAxisProps={{
                     scale: scaleLinear(),
                     keys: ['yes', 'no', 'maybe'],
                     title: 'lb/acre',
-                    titlePadding: 30
+                    titleYPadding: 30
                 }}
                 barFill={(values) => FIXTURE.colors[values.key]}
                 data={FIXTURE.data}