+++ gd-geostreaming/src/containers/Sensor/Detail/LineChartWrapper.jsx
@@ -36,8 +36,10 @@ const LineChartWrapper = (props: Props) => {
         tooltipContainerRef
     } = props;
 
-    const [lineChartContainer, lineChartContainerRect] = useElementRect();
-    const [boxPlotContainer, boxPlotContainerRect] = useElementRect();
+    const lineChartContainer = React.useRef();
+    const lineChartContainerRect = useElementRect(lineChartContainer);
+    const boxPlotContainer = React.useRef();
+    const boxPlotContainerRect = useElementRect(boxPlotContainer);
 
     const lineData = [];
     const boxData = [];
@@ -87,7 +89,7 @@ const LineChartWrapper = (props: Props) => {
                                 key: 'average',
                                 scale: scaleLinear().domain( [startAtZero ? 0 : minAverage, maxAverage]),
                                 title: unit,
-                                titlePadding: 40
+                                titlePadding: 10
                             }}
                             lineStroke="#56B4E9"
                             lineStrokeWidth={2}